+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * :bug:`7488` Show tags in multi-lines when multiple to avoid horizontal scroll.
 * :bug:`7497` In ETH staking view execution rewards should now be counted properly. MEV reward and block reward should not both be counted if recipient is not tracked.
 * :bug:`7522` Invalid ENS names shouldn't stop the decoding process anymore.
+* :bug:`-` Removing an address from one of the EVM chains won't affect the decoded events in other chains anymore.
 * :bug:`-` ETH withdrawal events should now be taxable again if the setting for their treatment after withdrawals enabled is on (which is by default).
 * :bug:`-` Invalid data in airdrops' CSVs or JSONs will now get ignored to show the rest of the valid data.